+++ src/hooks/common/Utility.hpp
@@ -19,6 +19,7 @@
 #include <rapidjson/document.h>
 
 #include "common/Logger.hpp"
+#include "common/UserIdentity.hpp"
 
 namespace sarus {
 namespace hooks {
@@ -29,6 +30,8 @@ void applyLoggingConfigIfAvailable(const rapidjson::Document&);
 std::tuple<boost::filesystem::path, pid_t> parseStateOfContainerFromStdin();
 std::unordered_map<std::string, std::string> parseEnvironmentVariablesFromOCIBundle(const boost::filesystem::path&);
 void enterNamespacesOfProcess(pid_t);
+void validatedBindMount(const boost::filesystem::path& from, const boost::filesystem::path& to,
+        const sarus::common::UserIdentity& userIdentity, const boost::filesystem::path& bundleDir, const boost::filesystem::path& rootfsDir);
 std::tuple<boost::filesystem::path, boost::filesystem::path> findSubsystemMountPaths(const std::string& subsystemName,
         const boost::filesystem::path& procPrefixDir, const pid_t pid);
 boost::filesystem::path findCgroupPathInHierarchy(const std::string& subsystemName, const boost::filesystem::path& procPrefixDir,